refactor(nested-collapsing-view): moves all classes from test package
