+++ nested-collapsing-view/src/main/java/com/samagra/nested_collapsing_view/HierarchyViewAdapter.java
@@ -1,4 +1,4 @@
-package com.samagra.nested_collapsing_view.test;
+package com.samagra.nested_collapsing_view;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -6,8 +6,6 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-
-import com.samagra.nested_collapsing_view.R;
 
 import java.util.ArrayList;
 import java.util.HashMap;
